requirements.txt
python-dotenv
ccxt
networkx
